Update for Open Position and Archived Closed Position
Update for Open Position and Archived Closed Position to return all data in ark_portfolio since data might have different dates because of below:

"So for top opens we should show the newest date that was part of the aggregation. So if it finds a few buys and a few sells when it does the aggregate, it should pick the date with the newest buy/sell.
For archived closes, it should use the NC (newest close) date"